Use CDN assets and externalize API URLs
