+++ Aspx/MxFace.Fingerprint.Aspx.Client/App_Code/Constants.cs
@@ -1,0 +1,7 @@
+namespace MxFace.Fingerprint.Aspx.Client.App_Code
+{
+    public static class Constants
+    {
+        public const string ApplicationName = "MyWebApp";
+    }
+}
