+++ Aspx/MxFace.Fingerprint.Aspx.Client/App_Themes/Default/SkinFile.skin
@@ -1,0 +1,1 @@
+<!-- Sample skin file -->
