+++ Aspx/MxFace.Fingerprint.Aspx.Client/App_Themes/Default/Styles.css
@@ -1,0 +1,2 @@
+/* Theme styles */
+body { font-family: Arial, sans-serif; }
