+++ Aspx/MxFace.Fingerprint.Aspx.Client/Global.asax
@@ -1,0 +1,1 @@
+<%@ Application Language="C#" Codebehind="Global.asax.cs" Inherits="MxFace.Fingerprint.Aspx.Client.Global" %>
