+++ Aspx/MxFace.Fingerprint.Aspx.Client/Global.asax.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace MxFace.Fingerprint.Aspx.Client
+{
+    public class Global : System.Web.HttpApplication
+    {
+        protected void Application_Start(object sender, EventArgs e)
+        {
+        }
+    }
+}
