+++ Aspx/MxFace.Fingerprint.Aspx.Client/Models/BaseResult.cs
