+++ Aspx/MxFace.Fingerprint.Aspx.Client/Models/CaptureViewModel.cs
