+++ Aspx/MxFace.Fingerprint.Aspx.Client/Models/Device.cs
