+++ Aspx/MxFace.Fingerprint.Aspx.Client/Services/FingerprintApiService.cs
@@ -5,17 +5,19 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 using System.Web.Script.Serialization;
+using System.Configuration;
 
 namespace MxFace.Fingerprint.Aspx.Client.Services
 {
     public class FingerprintApiService
     {
         private readonly HttpClient _httpClient;
-        private const string BaseUrl = "api/Fingerprint";
+        private readonly string _baseUrl;
 
         public FingerprintApiService(HttpClient httpClient)
         {
             _httpClient = httpClient;
+            _baseUrl = ConfigurationManager.AppSettings["FingerprintApiPath"];
         }
 
         public async Task<EnrollmentResponse> EnrollAsync(byte[] templateData, string personId, string group)
@@ -25,7 +27,7 @@ namespace MxFace.Fingerprint.Aspx.Client.Services
                 var serializer = new JavaScriptSerializer();
                 var request = new EnrollRequest { TemplateData = templateData, PersonId = personId, Group = group };
                 var content = new StringContent(serializer.Serialize(request), Encoding.UTF8, "application/json");
-                var response = await _httpClient.PostAsync(BaseUrl + "/enroll", content);
+                var response = await _httpClient.PostAsync(_baseUrl + "/enroll", content);
                 var body = await response.Content.ReadAsStringAsync();
                 if (response.IsSuccessStatusCode)
                 {
@@ -46,7 +48,7 @@ namespace MxFace.Fingerprint.Aspx.Client.Services
                 var serializer = new JavaScriptSerializer();
                 var request = new SearchRequest { TemplateData = templateData, Group = group };
                 var content = new StringContent(serializer.Serialize(request), Encoding.UTF8, "application/json");
-                var response = await _httpClient.PostAsync(BaseUrl + "/search", content);
+                var response = await _httpClient.PostAsync(_baseUrl + "/search", content);
                 var body = await response.Content.ReadAsStringAsync();
                 if (response.IsSuccessStatusCode)
                 {
@@ -67,7 +69,7 @@ namespace MxFace.Fingerprint.Aspx.Client.Services
                 var serializer = new JavaScriptSerializer();
                 var request = new MatchRequest { SourceTemplate = sourceTemplate, TargetTemplate = targetTemplate };
                 var content = new StringContent(serializer.Serialize(request), Encoding.UTF8, "application/json");
-                var response = await _httpClient.PostAsync(BaseUrl + "/verify", content);
+                var response = await _httpClient.PostAsync(_baseUrl + "/verify", content);
                 var body = await response.Content.ReadAsStringAsync();
                 if (response.IsSuccessStatusCode)
                 {
