+++ Aspx/MxFace.Fingerprint.Aspx.Client/Web.config
@@ -1,5 +1,10 @@
 <?xml version="1.0"?>
 <configuration>
+  <appSettings>
+    <add key="ApiBaseUrl" value="https://localhost:7103/" />
+    <add key="FingerprintApiPath" value="api/Fingerprint" />
+    <add key="ScanBaseUrl" value="https://localhost:8034/mfscan" />
+  </appSettings>
   <system.web>
     <compilation debug="true" targetFramework="4.8" />
     <httpRuntime targetFramework="4.8" />
